requirements.txt
Django==5.0.4
djangorestframework==3.15.1
django-environ==0.11.2
psycopg[binary]
Pillow==10.3.0
djangorestframework-simplejwt==5.3.1
